+++ foundation_stdlib/lib/matrix.lua
@@ -1,0 +1,2 @@
+foundation_stdlib:require("lib/matrix/3x3.lua")
+foundation_stdlib:require("lib/matrix/4x4.lua")