+++ foundation_stdlib/lib/quaternion.lua
@@ -1,0 +1,9 @@
+--- @namespace foundation.com.Quaternion
+
+local m = {}
+
+function m.new()
+  return {}
+end
+
+foundation.com.Quaternion = m