+++ foundation_stdlib/tests/matrix_test.lua
@@ -1,0 +1,4 @@
+local mod = assert(foundation_stdlib)
+
+mod:require("tests/matrix/3x3_test.lua")
+mod:require("tests/matrix/4x4_test.lua")